requirements.txt
selenium-wire==5.1.0
python-telegram-bot==20.4
webdriver_manager==4.0.0
python-dotenv==1.0.0
undetected-chromedriver==3.5.3
